journal: new generic journal implementation
The initial use-case for this journal is for librbd and its mirroring
feature. This journal is different from the current MDS journal in that
it organizes journal entries across journal objects w/o striping
individual entries.  It also allows multiple clients to read and write
from a single journal concurrently.

Signed-off-by: Jason Dillaman <dillaman@redhat.com>